flask_app/requirements.txt
Flask==2.3.1
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
statsmodels==0.13.5